# backend/api/openapi-v1.yaml
openapi: "3.1.0"
info:
  title: Detective Game API
  description: API для мобильной игры в жанре детективного расследования.
  version: "1.0.0"
servers:
  - url: http://localhost:8080
    description: Локальный сервер для разработки
security:
  - UserID: []
  - SessionID: []
paths:
  /api/v1/sessions:
    post:
      operationId: createSession
      tags: [Sessions]
      summary: Создать новую игровую сессию
      description: Запускает конвейер генерации сценария (Crime → Timeline → Evidence → Memories → Consistency Check). Сессия привязывается к пользователю.
      security:
        - UserID: []
      responses:
        "201":
          description: Сессия создана, генерация завершена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateSessionResponse"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/sessions/history:
    get:
      operationId: listHistory
      tags: [Sessions]
      summary: Список завершённых сессий
      description: Возвращает все предыдущие сессии пользователя (с результатами и без).
      security:
        - UserID: []
      responses:
        "200":
          description: Список сессий
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Session"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/sessions/current:
    get:
      operationId: getSession
      tags: [Sessions]
      summary: Текущая активная сессия
      description: Возвращает активную сессию пользователя (фаза не finished). Не требует X-Session-ID — ищет по X-User-ID.
      security:
        - UserID: []
      responses:
        "200":
          description: Активная сессия
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/sessions/{id}:
    get:
      operationId: getSessionById
      tags: [Sessions]
      summary: Конкретная сессия по ID
      description: Возвращает любую сессию пользователя по ID (только если принадлежит пользователю).
      security:
        - UserID: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID сессии
      responses:
        "200":
          description: Сессия
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/evidence:
    get:
      operationId: listEvidence
      tags: [Evidence]
      summary: Список улик
      description: Все вещественные улики, выданные в начале сессии.
      responses:
        "200":
          description: Список улик
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Evidence"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/evidence/{evId}:
    get:
      operationId: getEvidence
      tags: [Evidence]
      summary: Детали улики
      parameters:
        - $ref: "#/components/parameters/EvidenceId"
      responses:
        "200":
          description: Детальное описание улики
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Evidence"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/characters:
    get:
      operationId: listCharacters
      tags: [Characters]
      summary: Список персонажей в сессии
      description: Возвращает 5 персонажей текущей сессии с их знаниями, секретами, воспоминаниями и текущим уровнем доверия.
      responses:
        "200":
          description: Список персонажей
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Character"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/characters/{charId}:
    get:
      operationId: getCharacter
      tags: [Characters]
      summary: Детали персонажа
      parameters:
        - $ref: "#/components/parameters/CharacterId"
      responses:
        "200":
          description: Полная информация о персонаже
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Character"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/chronology:
    get:
      operationId: getChronology
      tags: [Chronology]
      summary: Хронология расследования
      description: "Лента всех событий в деле: допросы, результаты анализов, заметки."
      responses:
        "200":
          description: Хронология
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChronologyEntry"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/chronology/{chronId}/notes/{noteId}:
    patch:
      operationId: updateNotebookEntry
      tags: [Chronology]
      summary: Обновить запись в блокноте
      description: Пометить запись тегами (strange/suspicious/lie/key) и/или добавить заметку.
      parameters:
        - $ref: "#/components/parameters/ChronologyId"
        - $ref: "#/components/parameters/NoteId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tags:
                  type: array
                  items:
                    $ref: "#/components/schemas/NoteTag"
                  description: Метки для записи
                note:
                  type: string
                  description: Текстовая заметка игрока
      responses:
        "200":
          description: Запись обновлена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusOk"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/interrogations:
    post:
      operationId: createInterrogation
      tags: [Interrogations]
      summary: Начать допрос персонажа
      description: Создаёт новую сессию допроса указанного персонажа. У персонажа должен быть ненулевой лимит допросов.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - character_id
              properties:
                character_id:
                  type: string
                  format: uuid
                  description: ID персонажа для допроса
      responses:
        "201":
          description: Допрос создан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Interrogation"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/interrogations/{interId}:
    get:
      operationId: getInterrogation
      tags: [Interrogations]
      summary: Информация о допросе
      parameters:
        - $ref: "#/components/parameters/InterrogationId"
      responses:
        "200":
          description: Допрос
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Interrogation"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/interrogations/{interId}/messages:
    get:
      operationId: getInterrogationMessages
      tags: [Interrogations]
      summary: История сообщений допроса
      description: Полная история диалога с персонажем в рамках этого допроса.
      parameters:
        - $ref: "#/components/parameters/InterrogationId"
      responses:
        "200":
          description: Сообщения
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      operationId: addInterrogationMessage
      tags: [Interrogations]
      summary: Отправить сообщение в допросе
      description: Отправляет реплику игрока и получает ответ персонажа от LLM. LLM получает полную историю диалога (включая предыдущие допросы) и CharacterState.
      parameters:
        - $ref: "#/components/parameters/InterrogationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - message
              properties:
                message:
                  type: string
                  description: Текст вопроса/реплики игрока
      responses:
        "200":
          description: Ответ персонажа (сообщение от LLM)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/interrogations/{interId}/complete:
    patch:
      operationId: completeInterrogation
      tags: [Interrogations]
      summary: Завершить допрос
      description: Принудительно завершает допрос (или по истечении 10 минут, или по желанию игрока).
      parameters:
        - $ref: "#/components/parameters/InterrogationId"
      responses:
        "200":
          description: Допрос завершён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusOk"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/actions/dna-analysis:
    post:
      operationId: dnaAnalysis
      tags: [Actions]
      summary: Анализ ДНК
      description: Тратит 1 очко действия на анализ ДНК улики.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - evidence_id
              properties:
                evidence_id:
                  type: string
                  format: uuid
                  description: ID улики для анализа
      responses:
        "200":
          description: Результат анализа
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionReport"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/actions/fingerprints:
    post:
      operationId: fingerprintsCheck
      tags: [Actions]
      summary: Поиск отпечатков
      description: Тратит 1 очко действия на проверку отпечатков на улике.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - evidence_id
              properties:
                evidence_id:
                  type: string
                  format: uuid
                  description: ID улики для проверки
      responses:
        "200":
          description: Результат проверки
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionReport"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/actions/alibi-check:
    post:
      operationId: alibiCheck
      tags: [Actions]
      summary: Проверка алиби
      description: Тратит 1 очко действия на проверку алиби персонажа.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - character_id
                - alibi_text
              properties:
                character_id:
                  type: string
                  format: uuid
                  description: ID персонажа
                alibi_text:
                  type: string
                  description: Текст алиби для проверки
      responses:
        "200":
          description: Результат проверки алиби
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionReport"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/actions/camera-review:
    post:
      operationId: cameraReview
      tags: [Actions]
      summary: Запрос записей с камер
      description: Тратит 1 очко действия на запрос записей с камер наблюдения.
      responses:
        "200":
          description: Записи с камер
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionReport"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/actions/call-history:
    post:
      operationId: callHistory
      tags: [Actions]
      summary: Запрос истории звонков
      description: Тратит 1 очко действия на запрос телефонных записей персонажа.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - character_id
              properties:
                character_id:
                  type: string
                  format: uuid
                  description: ID персонажа
      responses:
        "200":
          description: История звонков
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionReport"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/actions/transactions:
    post:
      operationId: transactionCheck
      tags: [Actions]
      summary: Запрос банковских транзакций
      description: Тратит 1 очко действия на запрос финансовых операций персонажа.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - character_id
              properties:
                character_id:
                  type: string
                  format: uuid
                  description: ID персонажа
      responses:
        "200":
          description: Банковские транзакции
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionReport"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/reports/{reportId}:
    get:
      operationId: getReport
      tags: [Reports]
      summary: Получить отчёт
      description: Возвращает отчёт лаборатории (результат анализа или проверки).
      parameters:
        - $ref: "#/components/parameters/ReportId"
      responses:
        "200":
          description: Отчёт
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionReport"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/reports:
    post:
      operationId: submitReport
      tags: [Reports]
      summary: Сдать финальный отчёт
      description: |
        Игрок отправляет свою версию событий. Игра сравнивает с ground truth по 5 пунктам (who/why/how/when/evidence) — бинарно correct/incorrect.
        LLM генерирует нарративный фидбек.
        После этого сессия считается завершённой.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitReportRequest"
      responses:
        "200":
          description: Результат оценки
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
components:
  securitySchemes:
    UserID:
      type: apiKey
      in: header
      name: X-User-ID
      description: UUID пользователя (обязателен для всех запросов)
    SessionID:
      type: apiKey
      in: header
      name: X-Session-ID
      description: UUID активной сессии (обязателен для всех запросов кроме `POST /sessions`, `GET /sessions/history`, `GET /sessions/current` и `GET /sessions/{id}`)
  parameters:
    EvidenceId:
      name: evId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: ID улики
    CharacterId:
      name: charId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: ID персонажа
...
  schemas:
...
    CreateSessionResponse:
      type: object
      properties:
        session_id:
          type: string
          format: uuid
          description: ID созданной сессии
      required:
        - session_id
    Session:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        user_id:
          $ref: "#/components/schemas/UUID"
        case_name:
          type: string
          description: Название дела (генерируется LLM)
        case_brief:
          type: string
          description: Markdown-документ с описанием дела (для вкладки «Факты»)
        action_points:
          type: integer
          description: Оставшиеся очки действий (лупы)
          minimum: 0
          maximum: 5
        phase:
          type: string
          description: Текущая фаза сессии
          enum:
            - investigating
            - finished
        game_result:
          $ref: "#/components/schemas/GameResult"
        created_at:
          type: string
          format: date-time
        finished_at:
          type: string
          format: date-time
          description: Время завершения расследования
        content_locale:
          type: string
          description: Язык контента расследования, закреплённый при создании сессии
          enum:
            - ru
            - en
      required:
        - id
        - user_id
        - action_points
        - phase
        - created_at
        - content_locale
...
  responses:
    BadRequest:
      description: Невалидный запрос (некорректный JSON, отсутствуют обязательные поля, невалидный UUID)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "missing or invalid X-User-ID header"
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "not_found"
...
    InternalError:
      description: Внутренняя ошибка сервера
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "failed to create session"
...